+++ src/sagas/index.js
@@ -55,7 +55,7 @@ export default function* rootSaga({
   yield spawn(notificationSaga, { mobileApi });
   yield spawn(passwordSaga, { mobileApi, sessionStore, Crypto });
   yield spawn(permissionSaga, { permissionService });
-  yield spawn(plipSaga, { apiError, localStorage, mobileApi, walletStore });
+  yield spawn(plipSaga, { apiError, DeviceInfo, localStorage, mobileApi, walletStore });
   yield spawn(profileSaga, { dispatch, mobileApi, DeviceInfo, sessionStore, Crypto, walletStore });
   yield spawn(stateSaga, { repositories });
   yield spawn(errorSaga);