+++ src/services/local-storage.js
@@ -15,7 +15,17 @@ const fetch = (namespace, key) =>
 const destroy = (namespace, key) =>
   AsyncStorage.removeItem(buildKey(namespace, key));
 
+const findOrCreate = curry((namespace, key, valueBuilder) =>
+  fetch(namespace, key)
+  .then(async persistedValue => {
+    if (persistedValue) return persistedValue;
+
+    const value = await Promise.resolve().then(() => valueBuilder());
+    return store(namespace, key, value).then(() => value);
+  }));
+
 const service = namespace => ({
+  findOrCreate: findOrCreate(namespace),
   store: curry(store)(namespace),
   fetch: curry(fetch)(namespace),
   destroy: curry(destroy)(namespace),